requirements.txt
bleach==2.0.0
click==6.7
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
Flask==0.12.1
grpcio==1.2.1
grpcio-tools==1.2.1
html5lib==0.999999999
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==0.23
numpy==1.12.0
opencv-python==3.2.0.6
pickleshare==0.7.4
prompt-toolkit==1.0.13
protobuf==3.2.0
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
scipy==0.19.0
six==1.10.0
tensorflow==1.0.1
terminado==0.6
testpath==0.3
Theano==0.8.2
tornado==4.4.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
